compartmentalized Checkbox Dropdown Button feature for reusability
